Extract MainThreadWebviewSerializers to own file
Fixes #105127

This time make sure to avoid cycles too
